感情スコアグラフのコンポーネント化対応
